refactor: Restructure files and folders to match features
