+++ src/features/autosuggest/autosuggest.ts
@@ -1,15 +1,12 @@
 import { getConfig } from '../../shared/config';
 import { SUGGEST_ENDPOINT_PROD } from '../../shared/constants';
-import type { AccountParameters } from '../../shared/types/account-params.type';
-import type { CallTypeParameters } from '../../shared/types/calltype-params.type';
-import type { CatalogParameters } from '../../shared/types/catalog-params.type';
-import type { SuggestRequestParameters } from '../../shared/types/suggest-request.type';
-import type { SuggestResponse } from '../../shared/types/suggest-response.type';
 import { buildApiUrl } from '../../shared/url-builders';
+import type { SuggestRequestParameters } from './suggest-request.type';
+import type { SuggestResponse } from './suggest-response.type';
 
 export type AutosuggestOptions = Omit<
   SuggestRequestParameters,
-  keyof AccountParameters | keyof CatalogParameters | keyof CallTypeParameters
+  SuggestRequestParameters['account_id'] | SuggestRequestParameters['request_type']
 >;
 
 export async function autoSuggest(params: AutosuggestOptions): Promise<SuggestResponse> {
