+++ src/features/autosuggest/suggest-response.type.ts
