+++ src/features/search/bestseller/bestseller.spec.ts
@@ -1,10 +1,10 @@
 import { HttpResponse, http } from 'msw';
 import { describe, expect, test } from 'vitest';
-import { createBestsellerOptionsMock } from '../../shared/mocks/bestseller-options.mock';
-import { createSetupConfigMock } from '../../shared/mocks/configuration.mock';
-import { mockRequest } from '../../shared/mocks/mock-request.mock';
-import { createSearchResponseMock } from '../../shared/mocks/search-response.mock';
+import { createBestsellerOptionsMock } from './bestseller-options.mock';
 import { bestseller } from './bestseller';
+import { createSetupConfigMock } from '../../../initialize/configuration.mock';
+import { mockRequest } from '../../../shared/mock-request.mock';
+import { createSearchResponseMock } from '../search-response.mock';
 
 describe('Bestseller API', () => {
   const config = createSetupConfigMock();
