+++ src/features/search/content-search/content-search-options.mock.ts
@@ -1,17 +1,19 @@
 import { faker } from '@faker-js/faker';
-import type { BestsellerOptions } from '../../features/bestseller/bestseller';
+import type { ContentSearchOptions } from './content-search';
 
 /*
  * Create mock object with the option to override any of the generated fake values
  */
-export function createBestsellerOptionsMock(
-  overrides?: Partial<BestsellerOptions>,
-): BestsellerOptions {
-  const defaults: BestsellerOptions = {
+export function createContentSearchOptionsMock(
+  overrides?: Partial<ContentSearchOptions>,
+): ContentSearchOptions {
+  const defaults = {
+    catalog_name: faker.commerce.department(),
     q: faker.commerce.productName(),
     fl: [`pid`, Array.from({ length: 4 }, () => faker.database.column())].join(','),
     start: 0,
     rows: faker.number.int({ min: 1, max: 5 }),
+    'facet.version': 3.0,
     'facet.range': faker.database.column(),
     fq: `${faker.database.column()}: ${faker.lorem.word()}`,
     efq: `${faker.database.column()}: ${faker.lorem.word()}`,
@@ -21,8 +23,10 @@ export function createBestsellerOptionsMock(
     groupby: faker.database.column(),
     group_limit: faker.number.int({ min: 1, max: 5 }),
     group_offset: 0,
+    request_type: 'search',
+    search_type: 'keyword',
     _br_uid_2: faker.string.uuid(),
-    request_id: Number(faker.string.numeric(13)),
+    request_id: faker.string.numeric(13),
     url: faker.internet.url(),
     ref_url: faker.internet.url(),
     user_id: faker.string.uuid(),
@@ -30,7 +34,7 @@ export function createBestsellerOptionsMock(
     'query.precision': faker.helpers.arrayElement([
       'text_match_precision',
       'product_type_precision',
-      'category_precision',
+      'content_precision',
       'llm_based_precision',
     ]),
     'query.relaxation': Math.random() ? 'product_type' : 'off',
