+++ src/features/search/product-search/product-search-options.mock.ts
@@ -1,19 +1,17 @@
 import { faker } from '@faker-js/faker';
-import type { ContentSearchOptions } from '../../features/content-search/content-search';
+import type { ProductSearchOptions } from './product-search';
 
 /*
  * Create mock object with the option to override any of the generated fake values
  */
-export function createContentSearchOptionsMock(
-  overrides?: Partial<ContentSearchOptions>,
-): ContentSearchOptions {
-  const defaults = {
-    catalog_name: faker.commerce.department(),
+export function createProductSearchOptionsMock(
+  overrides?: Partial<ProductSearchOptions>,
+): ProductSearchOptions {
+  const defaults: ProductSearchOptions = {
     q: faker.commerce.productName(),
     fl: [`pid`, Array.from({ length: 4 }, () => faker.database.column())].join(','),
     start: 0,
     rows: faker.number.int({ min: 1, max: 5 }),
-    'facet.version': 3.0,
     'facet.range': faker.database.column(),
     fq: `${faker.database.column()}: ${faker.lorem.word()}`,
     efq: `${faker.database.column()}: ${faker.lorem.word()}`,
@@ -23,10 +21,8 @@ export function createContentSearchOptionsMock(
     groupby: faker.database.column(),
     group_limit: faker.number.int({ min: 1, max: 5 }),
     group_offset: 0,
-    request_type: 'search',
-    search_type: 'keyword',
     _br_uid_2: faker.string.uuid(),
-    request_id: faker.string.numeric(13),
+    request_id: Number(faker.string.numeric(13)),
     url: faker.internet.url(),
     ref_url: faker.internet.url(),
     user_id: faker.string.uuid(),
@@ -34,7 +30,7 @@ export function createContentSearchOptionsMock(
     'query.precision': faker.helpers.arrayElement([
       'text_match_precision',
       'product_type_precision',
-      'content_precision',
+      'category_precision',
       'llm_based_precision',
     ]),
     'query.relaxation': Math.random() ? 'product_type' : 'off',
