+++ src/features/search/search-response.type.ts
