+++ src/index.ts
@@ -1,14 +1,14 @@
 export { autoSuggest } from './features/autosuggest/autosuggest';
 export type { AutosuggestOptions } from './features/autosuggest/autosuggest';
-export { bestseller } from './features/bestseller/bestseller';
-export type { BestsellerOptions } from './features/bestseller/bestseller';
-export { categorySearch } from './features/category-search/category-search';
-export type { CategorySearchOptions } from './features/category-search/category-search';
-export { contentSearch } from './features/content-search/content-search';
-export type { ContentSearchOptions } from './features/content-search/content-search';
-export { productSearch } from './features/product-search/product-search';
-export type { ProductSearchOptions } from './features/product-search/product-search';
+export type { SuggestResponse } from './features/autosuggest/suggest-response.type';
+export type { SearchResponse } from './features/search/search-response.type';
+export { bestseller } from './features/search/bestseller/bestseller';
+export type { BestsellerOptions } from './features/search/bestseller/bestseller';
+export { categorySearch } from './features/search/category-search/category-search';
+export type { CategorySearchOptions } from './features/search/category-search/category-search';
+export { contentSearch } from './features/search/content-search/content-search';
+export type { ContentSearchOptions } from './features/search/content-search/content-search';
+export { productSearch } from './features/search/product-search/product-search';
+export type { ProductSearchOptions } from './features/search/product-search/product-search';
 export { initialize } from './initialize/initialize';
-export type { SetupConfiguration } from './shared/types/configuration.type';
-export type { SearchResponse } from './shared/types/search-response.type';
-export type { SuggestResponse } from './shared/types/suggest-response.type';
+export type { SetupConfiguration } from './initialize/configuration.type';
