--- src/shared/types/account-params.type.ts
@@ -1,7 +1,0 @@
-import type { account_id, auth_key } from './parameters';
-
-export interface AccountParameters {
-  account_id: account_id;
-  auth_key?: auth_key;
-}
-
