--- src/shared/types/calltype-params.type.ts
@@ -1,8 +1,0 @@
-import type { request_type, search_type } from './parameters';
-
-export interface CallTypeParameters {
-  request_type: request_type;
-  search_type: search_type;
-}
-
-
