--- src/shared/types/catalog-params.type.ts
@@ -1,7 +1,0 @@
-import type { domain_key, view_id } from './parameters';
-
-export interface CatalogParameters {
-  domain_key: domain_key;
-  view_id?: view_id;
-}
-
