--- src/shared/types/content-params.type.ts
@@ -1,5 +1,0 @@
-import type { catalog_name } from './parameters';
-
-export interface ContentParameters {
-  catalog_name: catalog_name;
-}
