--- src/shared/types/cors-params.type.ts
@@ -1,7 +1,0 @@
-import type { callback } from './parameters';
-
-export interface CORSParameters {
-  callback?: callback;
-}
-
-
