--- src/shared/types/tracking-params.type.ts
@@ -1,10 +1,0 @@
-import type { _br_uid_2, ref_url, request_id, url, user_id } from './parameters';
-
-export interface TrackingParameters {
-  _br_uid_2: _br_uid_2;
-  request_id?: request_id;
-  url: url;
-  ref_url?: ref_url;
-  user_id?: user_id;
-}
-
